+++ apollo-workspace/apollo/src/main/java/it/univaq/disim/mwt/apollo/business/impl/SurveyServiceImpl.java
@@ -135,38 +135,41 @@ public class SurveyServiceImpl implements SurveyService {
 
     @Override
     public void updateSurvey(Survey survey, MultipartFile file) throws BusinessException {
-        try {
-            if (file != null && !file.isEmpty()) {
-                if (!file.getContentType().equals("image/png") && !file.getContentType().equals("image/jpeg")) {
-                    throw new BusinessException("File format not valid: " + file.getContentType());
+        if(!survey.isActive()) {
+            try {
+                if (file != null && !file.isEmpty()) {
+                    if (!file.getContentType().equals("image/png") && !file.getContentType().equals("image/jpeg")) {
+                        throw new BusinessException("File format not valid: " + file.getContentType());
+                    }
+                    DocumentFile icon = ConversionUtility.multipartFile2DocumentFile(file);
+                    if (survey.getIcon() != null) {
+                        // Delete old icon
+                        documentFileService.delete(survey.getIcon());
+                    }
+                    // Set new icon
+                    documentFileService.create(icon);
+                    survey.setIcon(icon);
                 }
-                DocumentFile icon = ConversionUtility.multipartFile2DocumentFile(file);
-                if (survey.getIcon() != null) {
-                    // Delete old icon
-                    documentFileService.delete(survey.getIcon());
-                }
-                // Set new icon
-                documentFileService.create(icon);
-                survey.setIcon(icon);
-            }
 
-            surveyRepository.save(survey);
-        } catch (IOException | DataAccessException e) {
-            throw new BusinessException(e);
+                surveyRepository.save(survey);
+            } catch (IOException | DataAccessException e) {
+                throw new BusinessException(e);
+            }
         }
     }
 
     @Override
     public void deleteSurvey(Survey survey) throws BusinessException {
-
-        if (survey.getQuestionGroups() != null && survey.getQuestionGroups().size() > 0) {
-            Set<QuestionGroup> groups = survey.getQuestionGroups();
-            questionGroupService.deleteQuestionGroupList(groups);
+        if(!survey.isActive()) {
+            if (survey.getQuestionGroups() != null && survey.getQuestionGroups().size() > 0) {
+                Set<QuestionGroup> groups = survey.getQuestionGroups();
+                questionGroupService.deleteQuestionGroupList(groups);
+            }
+            if (survey.getInvitationPool() != null) {
+                invitationPoolService.deleteInvitationPool(survey.getInvitationPool());
+            }
+            surveyRepository.delete(survey);
         }
-        if (survey.getInvitationPool() != null) {
-            invitationPoolService.deleteInvitationPool(survey.getInvitationPool());
-        }
-        surveyRepository.delete(survey);
     }
 
 }
